feat(ritk-filter): AdditiveGaussianNoise bit-exact via ITK FastNorm port
Disproves the prior "RNG non-deterministic/impossible" verdict, which
assumed MersenneTwister. The Gaussian noise filter actually uses
itk::Statistics::NormalVariateGenerator (Marsaglia-MacLaren FastNorm),
which is deterministic. Ported it bit-for-bit (noise/fastnorm.rs: the
goto-based integer state machine -- m_Vec1 table rotations + chi-squared
renormalization + the LCG/SignedShiftXOR), seeded Hash(userSeed, 0) =
userSeed*2654435761, applied per voxel in scanline order.

Replaced the old rand+Box-Muller generator in noise/gaussian.rs (seed
u64 -> u32 to match sitk's uint32). Rust test asserts the exact sitk
noise sequence; Python differential matches sitk.AdditiveGaussianNoise
single-threaded (<=1e-3) across 3 std/mean/seed cases. Coverage
269 -> 270/298. SaltAndPepper/Shot/Speckle now also reachable (their
ITK generators are likewise deterministic).

+++ crates/ritk-python/src/filter/noise.rs
@@ -23,13 +23,13 @@ use std::sync::Arc;
 /// Returns:
 ///     PyImage with additive Gaussian noise applied.
 #[pyfunction]
-#[pyo3(signature = (image, std, mean=0.0, seed=42_u64))]
+#[pyo3(signature = (image, std, mean=0.0, seed=42_u32))]
 pub fn additive_gaussian_noise(
     py: Python<'_>,
     image: &PyImage,
     std: f64,
     mean: f64,
-    seed: u64,
+    seed: u32,
 ) -> RitkResult<PyImage> {
     let img = Arc::clone(&image.inner);
     let result = py
